+++ src/View/Helper/FormHelper.php
@@ -12,7 +12,6 @@ declare(strict_types=1);
  * @link        https://github.com/mirko-pagliai/me-tools
  * @license     https://opensource.org/licenses/mit-license.php MIT License
  */
-
 namespace MeTools\View\Helper;
 
 use Cake\Utility\Hash;