+++ tests/TestCase/Command/Install/FixComposerJsonCommandTest.php
@@ -86,6 +86,6 @@ class FixComposerJsonCommandTest extends TestCase
     {
         $this->exec($this->command . ' -p noExisting');
         $this->assertExitWithError();
-        $this->assertErrorContains('File or directory `noExisting` does not exist');
+        $this->assertErrorContains('File or directory `noExisting` is not writable');
     }
 }