+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # 2.x branch
 ## 2.21 branch
 ### 2.21.4
-* `append-text` and `prepend-text` options for `BootstrapFormHelper`
-    can handle buttons;
+* `append-text` and `prepend-text` options for `BootstrapFormHelper::control()`
+    method can handle buttons;
 * added `li()` and `meta()` methods for `BootstrapHtmlHelper`;
 * improved `Command::createDir()` method and message in case of `mkdir` error;
 * small and numerous improvements of descriptions, tags and code suggested by