+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 2.x branch
 ## 2.21 branch
+### 2.21.5
+* updated for php-tools 1.6.5.
+
 ### 2.21.4
 * `append-text` and `prepend-text` options for `BootstrapFormHelper::control()`
     method can handle buttons;